+++ meanfi/tb/__init__.py
@@ -2,7 +2,7 @@
 
 from .ops import add_tb, scale_tb
 from .transforms import ifftn_to_tb, kgrid_to_tb, tb_to_kfunc, tb_to_kgrid
-from .utils import fermi_energy, generate_tb_keys
+from .utils import generate_tb_keys
 
 __all__ = [
     "add_tb",
@@ -11,6 +11,5 @@ __all__ = [
     "kgrid_to_tb",
     "tb_to_kfunc",
     "tb_to_kgrid",
-    "fermi_energy",
     "generate_tb_keys",
 ]